メモリ改善

+++ maven-root/mtsa/src/main/java/ltsa/updatingControllers/stepwise/delayed/StepwiseDelayedEvaluationMetrics.java
@@ -28,7 +28,7 @@ final class StepwiseDelayedEvaluationMetrics {
             String label,
             Set<Integer> scope,
             int states,
-            int transitions,
+            long transitions,
             long countTime) {
         String scopedLabel = scopedMetricLabel(label, scope);
         UpdatingControllerEvaluationRecorder.recordCount(
